+++ spec/rubocop/cop/style/infinite_loop_spec.rb
@@ -1,0 +1,36 @@
+# encoding: utf-8
+
+require 'spec_helper'
+
+describe RuboCop::Cop::Style::InfiniteLoop do
+  subject(:cop) { described_class.new }
+
+  %w(1 2.0 [1] {}).each do |lit|
+    it "registers an offense for a while loop with #{lit} as condition" do
+      inspect_source(cop,
+                     ["while #{lit}",
+                      '  top',
+                      'end'
+                     ])
+      expect(cop.offenses.size).to eq(1)
+    end
+  end
+
+  %w(false nil).each do |lit|
+    it "registers an offense for a until loop with #{lit} as condition" do
+      inspect_source(cop,
+                     ["until #{lit}",
+                      '  top',
+                      'end'
+                     ])
+      expect(cop.offenses.size).to eq(1)
+    end
+  end
+
+  it 'accepts Kernel#loop' do
+    inspect_source(cop,
+                   ['loop { break if something }'])
+
+    expect(cop.offenses).to be_empty
+  end
+end